+++ IntegrationTests/LibraryCore.IntegrationTests.Kafka/Fixtures/WebApplicationFactoryFixture.cs
@@ -19,7 +19,7 @@ public class WebApplicationFactoryFixture : IDisposable
     }
 
 #if DEBUG
-    public const string SkipReason = "Don't want to run kafka integration tests locally";
+    public const string SkipReason = "";
 #else
     public const string SkipReason = "";
 #endif
